[FEAT] #40 Add ToastEntity, Divide Codes
